+++ src/main/java/roomescape/reservation/reservation/ReservationRepository.java
@@ -1,8 +1,8 @@
-package roomescape.user.repository.reservation;
+package roomescape.reservation.reservation;
 
 import java.util.List;
 import java.util.Optional;
-import roomescape.user.domain.Reservation;
+import roomescape.reservation.domain.Reservation;
 
 public interface ReservationRepository {
