+++ src/main/java/roomescape/time/controller/ReservationTimeRestController.java
@@ -1,4 +1,4 @@
-package roomescape.admin.controller;
+package roomescape.time.controller;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -10,10 +10,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.admin.controller.dto.ReservationTimeRequest;
-import roomescape.admin.controller.dto.ReservationTimeResponse;
-import roomescape.admin.domain.ReservationTime;
-import roomescape.admin.service.ReservationTimeService;
+import roomescape.time.controller.dto.ReservationTimeRequest;
+import roomescape.time.controller.dto.ReservationTimeResponse;
+import roomescape.time.domain.ReservationTime;
+import roomescape.time.service.ReservationTimeService;
 
 @RequiredArgsConstructor
 @RestController
