+++ src/test/java/roomescape/admin/repository/time/H2ReservationTimeRepositoryTest.java
@@ -9,7 +9,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
-import roomescape.admin.domain.ReservationTime;
+import roomescape.time.domain.ReservationTime;
+import roomescape.time.time.H2ReservationTimeRepository;
 
 @JdbcTest
 @Import(H2ReservationTimeRepository.class)
